ref(types) Split out some type modules
When we started on the Typescript conversion we used `types/index` as
a place to put all the types as we collected them. That module is now
quite sizable. We already have a few type modules and I think that is
a pattern we should continue moving towards.

This extracts a few (but not all) type modules out of `types/index`.
I've added re-exports from `types/index` so that using types doesn't
require developers to remember which module types come from as often.

If folks are happy with this concept and the implementation thus far
I'll continue and extract more type modules out.
